namespace log を廃止
