+++ src/lib/auto_close_fd.cpp
@@ -1,6 +1,6 @@
 #include "auto_close_fd.hpp"
 #include <unistd.h>
-#include "utils/log.hpp"
+#include "utils/logger.hpp"
 #include "utils/string.hpp"
 
 AutoCloseFd::AutoCloseFd(const int fd):
